python-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
httpx==0.26.0
Pillow==10.2.0
img2pdf==0.5.1
python-dotenv==1.0.0
pydantic==2.5.3
opencv-python-headless==4.9.0.80
numpy==1.24.4

# PDF manipulation
pikepdf>=8.0.0,<9.0.0

# OCR
pytesseract==0.3.10

# PDF text extraction
PyPDF2>=3.0.0

# PDF conversions
pdf2docx==0.5.8
pdf2image==1.17.0
python-pptx==0.6.23
python-docx==1.1.0
openpyxl==3.1.2

# PDF overlays (page numbers, watermarks, signatures)
reportlab>=3.6,<4.1

# Background removal (AI-based, U2Net)
rembg-serverless

# System monitoring
psutil==5.9.7

# Structured logging
python-json-logger==2.0.7

# Testing
pytest>=7.0
pytest-asyncio>=0.21.0
